// apps/web/src/lib/api/generated/main/model/lootlogConfigNpcResponseDtoOutput.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { LootlogConfigNpcResponseDtoOutputAllowedRaritiesItem } from './lootlogConfigNpcResponseDtoOutputAllowedRaritiesItem';
import type { LootlogConfigNpcResponseDtoOutputNpcType } from './lootlogConfigNpcResponseDtoOutputNpcType';

export interface LootlogConfigNpcResponseDtoOutput {
  id: number;
  npcType: LootlogConfigNpcResponseDtoOutputNpcType;
  allowedRarities: LootlogConfigNpcResponseDtoOutputAllowedRaritiesItem[];
}