// apps/web/src/lib/api/generated/main/model/lootlog-config-npc-response-dto-output.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { LootlogConfigNpcResponseDtoOutputAllowedRaritiesItem } from './lootlog-config-npc-response-dto-output-allowed-rarities-item.js';
import type { LootlogConfigNpcResponseDtoOutputNpcType } from './lootlog-config-npc-response-dto-output-npc-type.js';

export interface LootlogConfigNpcResponseDtoOutput {
  id: number;
  npcType: LootlogConfigNpcResponseDtoOutputNpcType;
  allowedRarities: LootlogConfigNpcResponseDtoOutputAllowedRaritiesItem[];
}