// apps/web/src/lib/api/generated/main/model/lootlog-config-npc-response-dto-output-allowed-rarities-item.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type LootlogConfigNpcResponseDtoOutputAllowedRaritiesItem = typeof LootlogConfigNpcResponseDtoOutputAllowedRaritiesItem[keyof typeof LootlogConfigNpcResponseDtoOutputAllowedRaritiesItem];


export const LootlogConfigNpcResponseDtoOutputAllowedRaritiesItem = {
  UNIQUE: 'UNIQUE',
  HEROIC: 'HEROIC',
  LEGENDARY: 'LEGENDARY',
  UPGRADED: 'UPGRADED',
} as const;