// apps/web/src/lib/api/generated/main/model/update-lootlog-config-npc-dto-allowed-rarities-item.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type UpdateLootlogConfigNpcDtoAllowedRaritiesItem = typeof UpdateLootlogConfigNpcDtoAllowedRaritiesItem[keyof typeof UpdateLootlogConfigNpcDtoAllowedRaritiesItem];


export const UpdateLootlogConfigNpcDtoAllowedRaritiesItem = {
  UNIQUE: 'UNIQUE',
  HEROIC: 'HEROIC',
  LEGENDARY: 'LEGENDARY',
  UPGRADED: 'UPGRADED',
} as const;